+++ matchbox-server/src/main/java/ch/ahdis/fhir/hapi/jpa/validation/ImplementationGuideProviderR4.java
@@ -60,6 +60,7 @@ import ca.uhn.fhir.rest.param.UriAndListParam;
 import ca.uhn.fhir.rest.server.SimpleBundleProvider;
 import ch.ahdis.matchbox.CliContext;
 import ch.ahdis.matchbox.MatchboxEngineSupport;
+import ch.ahdis.matchbox.engine.MatchboxEngine;
 import ch.ahdis.matchbox.engine.cli.VersionUtil;
 import ch.ahdis.matchbox.util.MatchboxPackageInstallerImpl;
 
@@ -104,12 +105,6 @@ public class ImplementationGuideProviderR4 extends ca.uhn.fhir.jpa.rp.r4.Impleme
 		OperationOutcome oo = load(theResource);
 		MethodOutcome outcome = new MethodOutcome();
 		outcome.setOperationOutcome(oo);
-
-		// initialize matchbox engine
-		log.info("Initializing matchbox engine(s): " + VersionUtil.getMemory());
-		matchboxEngineSupport.getMatchboxEngine(FHIRVersion._4_0_1.getDisplay(), this.cliContext, false, true);
-		log.info("Initializing matchbox engine finished: " + VersionUtil.getMemory());
-
 		return outcome;
 	}
 
@@ -198,6 +193,17 @@ public class ImplementationGuideProviderR4 extends ca.uhn.fhir.jpa.rp.r4.Impleme
 				.setName(theResource.getName())
 				.setVersion(theResource.getVersion())
 				.setInstallMode(PackageInstallationSpec.InstallModeEnum.STORE_ONLY));
+
+		if (cliContext!=null && cliContext.getOnlyOneEngine()) {
+			MatchboxEngine engine = matchboxEngineSupport.getMatchboxEngine(FHIRVersion._4_0_1.getDisplay(), this.cliContext, false, false);
+			try {
+				engine.loadPackage(theResource.getUrl(), theResource.getVersion());
+			} catch (Exception e) {
+				log.error("Error loading package " + theResource.getUrl() + " " + theResource.getVersion(), e);
+			}
+		} else {
+			matchboxEngineSupport.getMatchboxEngine(FHIRVersion._4_0_1.getDisplay(), this.cliContext, false, true);
+		}
 		return getOperationOutcome(installOutcome);
 	}
 
@@ -226,7 +232,18 @@ public class ImplementationGuideProviderR4 extends ca.uhn.fhir.jpa.rp.r4.Impleme
 		matchboxEngineSupport.setInitialized(true);
 		log.info("Initializing packages finished " + VersionUtil.getMemory());
 		log.info("Creating cached engines during startup  " + VersionUtil.getMemory());
-		matchboxEngineSupport.getMatchboxEngine(null,this.cliContext, false, false);
+		MatchboxEngine engine = matchboxEngineSupport.getMatchboxEngine(null,this.cliContext, false, true);
+		if (cliContext!=null && cliContext.getOnlyOneEngine()) {
+			List<NpmPackageVersionEntity> packages = myPackageVersionDao
+					.findAll(org.springframework.data.domain.Sort.by(Direction.ASC, "myPackageId", "myVersionId"));
+			for (NpmPackageVersionEntity npmPackage : packages) {
+				try {
+					engine.loadPackage(npmPackage.getPackageId(), npmPackage.getVersionId());
+				} catch (Exception e) {
+					log.error("Error loading package " + npmPackage.getPackageId() + " " + npmPackage.getVersionId(), e);
+				}
+			}
+		}
 		log.info("Finished engines during startup  " + VersionUtil.getMemory());
 		return installOutcome;
 	}