+++ matchbox-engine/src/main/java/org/hl7/fhir/r5/elementmodel/XmlParser.java
@@ -298,9 +298,6 @@ public class XmlParser extends ParserBase {
 		reapComments(node, element);
 
 		String xsiType = XMLUtil.getXsiType(node);
-		if (xsiType!=null && ("urn:hl7-org:v3".equals(node.getNamespaceURI()) || "urn:hl7-org:sdtc".equals(node.getNamespaceURI()) || "urn:ihe:pharm".equals(node.getNamespaceURI()) || "urn:oid:1.3.6.1.4.1.19376.1.3.2".equals(node.getNamespaceURI()))) {
-            xsiType = xsiType.toUpperCase(); // PATCH for italian CDA where ST is lowercase see ahdis/matchbox #145
-        }
 		List<Property> properties = element.getProperty().getChildProperties(element.getName(), xsiType);
 
 		String text = XMLUtil.getDirectText(node).trim();
@@ -309,22 +306,9 @@ public class XmlParser extends ParserBase {
 		if (!Utilities.noString(text)) {
 			Property property = getTextProp(properties);
 			if (property != null) {
-				if (("ST.data[x]".equals(property.getDefinition().getId()) || (property.getDefinition()!=null && property.getDefinition().getBase()!=null && "ST.data[x]".equals(property.getDefinition().getBase().getPath()))) 
-				 || ("ED.data[x]".equals(property.getDefinition().getId()) || (property.getDefinition()!=null && property.getDefinition().getBase()!=null && "ED.data[x]".equals(property.getDefinition().getBase().getPath())))){
-					if ("B64".equals(node.getAttribute("representation"))) {
-						Element n = new Element("data", property, "base64Binary", text).markLocation(line, col);
-						n.setPath(element.getPath()+"."+property.getName());
-						element.getChildren().add(n);
-					} else {
-						Element n = new Element("data", property, "string", text).markLocation(line, col);
-						n.setPath(element.getPath()+"."+property.getName());
-						element.getChildren().add(n);
-					}
-				} else {
-					Element n = new Element(property.getName(), property, property.getType(), text).markLocation(line, col);
-					n.setPath(element.getPath()+"."+property.getName());
-					element.getChildren().add(n);
-				}
+				Element n = new Element(property.getName(), property, property.getType(), text).markLocation(line, col);
+				n.setPath(element.getPath()+"."+property.getName());
+				element.getChildren().add(n);
 			}
 			else {
 				Node n = node.getFirstChild();
@@ -444,9 +428,6 @@ public class XmlParser extends ParserBase {
 								} else {
 									if (xsiType.contains(":"))
 										xsiType = xsiType.substring(xsiType.indexOf(":")+1);
-									if ("urn:hl7-org:v3".equals(node.getNamespaceURI()) || "urn:hl7-org:sdtc".equals(node.getNamespaceURI()) || "urn:ihe:pharm".equals(node.getNamespaceURI()) || "urn:oid:1.3.6.1.4.1.19376.1.3.2".equals(node.getNamespaceURI())) {
-										xsiType = xsiType.toUpperCase(); // PATCH for italian CDA where ST is lowercase see matchbox issue #145
-									}
 									n.setType(xsiType);
 									n.setExplicitType(xsiType);
 								}